Fix data overlow bug in maxMsgSize computation (#3782)
When we compute the 90% of this we were doing
an integer value of 64 MB times 90 which overflows
and results in a value of 17MB computed max size.
This limits the amount of data that is batched and has performance
ramifications.
fix long in BaseLogEntryReader & BaseSnapshotReader
NullPointerException fix in LogReplicationClientServerRouter

+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/logreader/BaseSnapshotReader.java
@@ -44,7 +44,7 @@ public abstract class BaseSnapshotReader extends SnapshotReader {
     /**
      * The max size of data for SMR entries in data message.
      */
-    protected final int maxDataSizePerMsg;
+    protected final long maxDataSizePerMsg;
     private final Optional<DistributionSummary> messageSizeDistributionSummary;
     protected final CorfuRuntime rt;
     protected long snapshotTimestamp;
